Move safe_string into delta
"There's nothing more permanent than a temporary fix"
