+++ engines/basic/include/delta_core.h
@@ -2,6 +2,5 @@
 #define DELTA_BASIC_DELTA_CORE_H
 
 #include "../../../include/yds_core.h"
-#include "../include/safe_string.h"
 
 #endif /* DELTA_BASIC_DELTA_CORE_H */
