+++ include/yds_safe_string.h
